Add Block dsl. (#11)

+++ Sources/Wayland/Text.swift
@@ -1,10 +1,10 @@
 public struct Text {
     public let text: String
-    public let pos: (x: Float, y: Float)
-    public let scale: Float
+    public let pos: (x: UInt, y: UInt)
+    public let scale: UInt
     public let color: Color
     public init(
-        _ text: String, at pos: (x: Float, y: Float), scale: Float, color: Color = .white
+        _ text: String, at pos: (x: UInt, y: UInt), scale: UInt, color: Color = .white
     ) {
         self.text = text
         self.pos = pos
